+++ src/main/java/top/offsetmonkey538/loottablemodifier/api/resource/util/LootModifierContext.java
@@ -1,0 +1,12 @@
+package top.offsetmonkey538.loottablemodifier.api.resource.util;
+
+import net.minecraft.loot.LootPool;
+import net.minecraft.loot.LootTable;
+import net.minecraft.loot.entry.LootPoolEntry;
+import net.minecraft.util.Identifier;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+
+public record LootModifierContext(@NotNull LootTable table, @NotNull Identifier tableId, @Nullable LootPool pool, @Nullable LootPoolEntry entry, boolean tableAlreadyModified, boolean poolAlreadyModified) {
+
+}
