+++ src/main/java/top/offsetmonkey538/loottablemodifier/datagen/LootTableModifierDatagen.java
@@ -10,7 +10,6 @@ import net.minecraft.loot.LootPool;
 import net.minecraft.loot.LootTable;
 import net.minecraft.loot.context.LootContextTypes;
 import net.minecraft.loot.entry.ItemEntry;
-import net.minecraft.loot.function.EnchantWithLevelsLootFunction;
 import net.minecraft.loot.function.SetCountLootFunction;
 import net.minecraft.loot.provider.number.ConstantLootNumberProvider;
 import net.minecraft.loot.provider.number.UniformLootNumberProvider;
@@ -18,13 +17,13 @@ import net.minecraft.registry.RegistryKey;
 import net.minecraft.registry.RegistryKeys;
 import net.minecraft.registry.RegistryWrapper;
 import top.offsetmonkey538.loottablemodifier.api.datagen.LootModifierProvider;
-import top.offsetmonkey538.loottablemodifier.resource.LootModifier;
-import top.offsetmonkey538.loottablemodifier.resource.OptionalPattern;
-import top.offsetmonkey538.loottablemodifier.resource.action.pool.AddPoolAction;
-import top.offsetmonkey538.loottablemodifier.resource.action.entry.SetItemAction;
-import top.offsetmonkey538.loottablemodifier.resource.action.pool.RemovePoolAction;
-import top.offsetmonkey538.loottablemodifier.resource.predicate.entry.ItemEntryPredicate;
-import top.offsetmonkey538.loottablemodifier.resource.predicate.table.LootTablePredicate;
+import top.offsetmonkey538.loottablemodifier.api.resource.LootModifier;
+import top.offsetmonkey538.loottablemodifier.api.resource.util.OptionalPattern;
+import top.offsetmonkey538.loottablemodifier.api.resource.action.pool.AddPoolAction;
+import top.offsetmonkey538.loottablemodifier.api.resource.action.entry.SetItemAction;
+import top.offsetmonkey538.loottablemodifier.api.resource.action.pool.RemovePoolAction;
+import top.offsetmonkey538.loottablemodifier.api.resource.predicate.entry.ItemEntryPredicate;
+import top.offsetmonkey538.loottablemodifier.api.resource.predicate.table.LootTablePredicate;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.BiConsumer;
